Restructured GML used by CityGML
Instead of using GML Owl classes from geosparql, GML 3.1.1 will be added to the ontology as implemented in the original CityGML
Owl. The reason for this change is geosparql's GML ontology is based off of GML 3.2 which does not adopt the same naming
conventions as GML 3.1.1. The approach moving forward will be to complete the GML 3.1.1 classes within the ontology, and use
the `owl:equivalentClass` tag to link those classes to geosparql's GML classes.

- reverted geometry classes to original CityGML owl
- restructured GML classes to respect GML 3.1.1 UML
- added missing GML classes and subclasses
- updated readme